move remaining resolveTypes to SDL
