+++ packages/server/graphql/public/typeDefs/Threadable.graphql
@@ -8,11 +8,6 @@ interface Threadable {
   id: ID!
 
   """
-  The rich text body of the item
-  """
-  content: String!
-
-  """
   The timestamp the item was created
   """
   createdAt: DateTime!
