+++ packages/server/graphql/types/Reactable.ts
@@ -1,7 +1,6 @@
 import {GraphQLID, GraphQLInterfaceType, GraphQLList, GraphQLNonNull} from 'graphql'
 import resolveReactjis from '../resolvers/resolveReactjis'
 import Reactji from './Reactji'
-import getReactableType from './getReactableType'
 
 export const reactableFields = () => ({
   id: {
@@ -17,17 +16,7 @@ export const reactableFields = () => ({
 
 const Reactable = new GraphQLInterfaceType({
   name: 'Reactable',
-  description: 'An item that can have reactjis',
-  fields: reactableFields,
-  resolveType: (type) => {
-    const reactableType = getReactableType(type)
-    const lookup = {
-      COMMENT: 'Comment',
-      REFLECTION: 'RetroReflection',
-      RESPONSE: 'TeamPromptResponse'
-    } as const
-    return lookup[reactableType]
-  }
+  fields: {}
 })
 
 export default Reactable
